requirements.txt
asgiref~=3.8
Django~=5.2
sqlparse==0.5.0
Pillow~=11.2
python-decouple==3.8
social-auth-app-django==5.4.0
django-extensions==3.2.3
werkzeug==3.0.2
pyOpenSSL==24.1.0
requests==2.31.0
easy-thumbnails==2.8.5